Move data buffer to blob buffer package
