+++ util/blob/buffer.go
@@ -1,4 +1,4 @@
-package data
+package blob
 
 import "math"
